Feat: category remove 전략 및 로직 작성
- 삭제 시 study_time 을 REMOVED 타입으로 변경, typeId 삭제, 단 현재 진행 중인 공부의 카테고리 변경 불가
- 목표 시간 저장 로직 추가

Ref: #97

+++ src/main/java/com/studypals/domain/studyManage/entity/StudyTime.java
@@ -56,7 +56,7 @@ public class StudyTime {
     private Long time = 0L;
 
     @Column(name = "goal", nullable = true)
-    private Integer goal;
+    private Long goal;
 
     public void addTime(Long time) {
         this.time += time;
